reorganized utility function files and created default exports
