+++ src/index.ts
@@ -1,2 +1,4 @@
 export * from "./components";
 export * from "./theme";
+export * from "./types";
+export * from "./utils";
