+++ src/types.ts
@@ -1,1 +1,2 @@
 export { RecursivePartial } from "./utils/deepMerge";
+export { Theme } from "@emotion/react";
