+++ src/utils/index.ts
@@ -1,0 +1,3 @@
+export { default as deepMerge } from "./deepMerge";
+export { default as shadeColor } from "./color/shadeColor";
+export { default as highestTextContrast } from "./color/highestTextContrast";
